\name{BMGMM-package}
\alias{BMGMM-package}
\alias{BMGMM}
\docType{package}
\title{
\packageTitle{BMGMM}
}
\description{
\packageDescription{BMGMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BMGMM}
\packageIndices{BMGMM}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{BMGMM}

Maintainer: \packageMaintainer{BMGMM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
